Update Gem lock for github-pages 94
It seems like you shouldn't put this in the repo since all I really
want is the latest of what's in the Gemfile (I assume Github just kind
of ignores everything anyway, they state the Pages gem is strict).

But it seems like it might be nice for something in the future. I
dunno, most Pages repos have it, and there's barely any Pages-specific
advice on it, so...